Update Project To iGap Doc Build 14

+++ app/src/main/java/com/iGap/activitys/ActivityMain.java
@@ -1175,7 +1175,7 @@ public class ActivityMain extends ActivityEnhanced implements OnComplete, OnChat
     }
 
     @Override
-    public void onChatUpdateStatus(long roomId, long messageId, ProtoGlobal.RoomMessageStatus status, int statusVersion) {
+    public void onChatUpdateStatus(long roomId, long messageId, ProtoGlobal.RoomMessageStatus status, long statusVersion) {
         // empty
     }
 }
