+++ app/src/main/java/com/iGap/interface_package/OnChatUpdateStatusResponse.java
@@ -2,9 +2,6 @@ package com.iGap.interface_package;
 
 import com.iGap.proto.ProtoGlobal;
 
-/**
- * Created by Alireza Eskandarpour Shoferi (meNESS) on 9/5/2016.
- */
 public interface OnChatUpdateStatusResponse {
-    void onChatUpdateStatus(long roomId, long messageId, ProtoGlobal.RoomMessageStatus status, int statusVersion);
+    void onChatUpdateStatus(long roomId, long messageId, ProtoGlobal.RoomMessageStatus status, long statusVersion);
 }
